app/requirements.txt
fastapi
uvicorn
pika
celery
sqlalchemy
psycopg2-binary